Create router to separate API routes and tested GET request to /api/task using a simple HTTP file.
