+++ http/server.http
@@ -1,0 +1,1 @@
+GET http://localhost:3000/api/task
